tentando arrumar bug master

+++ futurize-frontend/src/components/Table/Table.jsx
@@ -56,11 +56,9 @@ export default function TableC() {
 
     setOpen(true);
   };
-
   const handleClose = () => {
     setOpen(false);
   };
-
   const openDeleteConfirmation = (id) => {
     setIdToDelete(id);
     setDeleteConfirmationOpen(true);
@@ -77,11 +75,9 @@ export default function TableC() {
     // Feche a caixa de diálogo de confirmação
     setDeleteConfirmationOpen(false);
   };
-
   const handleEditClose = () => {
     setEditOpen(false);
   };
-
   const [formProjeto, setFormProjeto] = useState({
     titulo: '',
     inicio: '',
@@ -98,7 +94,6 @@ export default function TableC() {
       setFormProjeto({ ...formProjeto, [title]: value });
     }
   }
-
   function getStatusTagClass(estado) {
     switch (estado) {
       case 'ANDAMENTO':
@@ -111,7 +106,6 @@ export default function TableC() {
         return 'tag-status';
     }
   }
-
   useEffect(() => {
     // Função para buscar os dados do banco e preencher o estado 'rows' ao carregar a página
     const fetchData = async () => {
@@ -126,13 +120,11 @@ export default function TableC() {
         console.error("Erro ao conectar-se ao backend:", error);
       }
     };
-
     fetchData(); // Chame a função para buscar os dados ao carregar a página
   }, []);
 
   const handleFormSubmit = async (e) => {
     e.preventDefault();
-
 
     const { titulo, inicio, encerramento, estado, gestor } = formProjeto;
 
@@ -140,21 +132,19 @@ export default function TableC() {
     const dataInicial = inicio
       ? format(parse(inicio, 'dd-MM-yyyy', new Date()), 'yyyy-MM-dd')
       : format(new Date(), 'yyyy-MM-dd');
-
     const encerramentoEmData = parse(encerramento, 'dd-MM-yyyy', new Date());
-
     const dataFinal = format(encerramentoEmData, 'yyyy-MM-dd');
-
     const newRow = {
       titulo: titulo,
       inicio: dataInicial,
       encerramento: dataFinal,
       estado: estado.toUpperCase(),
-      gestor: getLoginUser(),
+      gestor: usuarioLogado,
     };
+    console.log(newRow);
 
     try {
-      const response = await axios.post(`http://localhost:8080/Porjeto/porUsuario/${usuarioLogado}`, newRow);
+      const response = await axios.post("http://localhost:8080/Projeto", newRow);
 
       if (response.status === 200) {
         const updatedRows = [...rows, newRow];
@@ -163,23 +153,19 @@ export default function TableC() {
         handleClose();
         return;
       }
-
       console.error('Erro ao salvar os dados no backend.');
     } catch (error) {
       console.error('Erro ao conectar-se ao backend:', error);
     }
   };
-
   const confirmDelete = async () => {
-    if (idToDelete !== undefined && !isNaN(idToDelete)) {
+    if (idToDelete !== null && idToDelete !== undefined && !isNaN(idToDelete)) {
       try {
         // Faça a chamada de exclusão para o backend
         await axios.delete(`http://localhost:8080/Projeto/${idToDelete}`);
-
         // Atualize o estado `rows` após a exclusão
         const updatedRows = rows.filter((row) => row.id !== idToDelete);
         setRows(updatedRows);
-
         // Feche a caixa de diálogo de confirmação
         setDeleteConfirmationOpen(false);
       } catch (error) {
@@ -189,7 +175,7 @@ export default function TableC() {
       console.error('ID de projeto inválido:', idToDelete);
     }
   };
-
+  
   // function isError() {
   //   handleClose();
   //   AlertError({
@@ -197,7 +183,6 @@ export default function TableC() {
   //     title: "Erro!",
   //   });
   // }
-
   const handleEditSubmit = async (e) => {
     e.preventDefault();
 
@@ -213,12 +198,10 @@ export default function TableC() {
 
     try {
       const response = await axios.put(`http://localhost:8080/Projeto/${id}`, updatedProjectData);
-
       if (response.status === 200) {
         // Atualize o estado `rows` após a edição
         const updatedRows = rows.map((row) => (row.id === editProjectData.id ? { ...row, ...updatedProjectData } : row));
         setRows(updatedRows);
-
         // Feche o modal de edição
         handleEditClose();
       } else {
@@ -228,7 +211,6 @@ export default function TableC() {
       console.error('Erro ao conectar-se ao backend:', error);
     }
   };
-
   const openEditProject = (project) => {
     // Preencha o estado `formProjeto` com os valores do projeto selecionado
     setFormProjeto({
@@ -238,17 +220,14 @@ export default function TableC() {
       encerramento: format(new Date(project.encerramento), 'yyyy-MM-dd'),
       estado: project.estado,
     });
-
     setEditProjectData(project);
     setEditOpen(true);
   };
-
   const openProjectKanban = (project) => {
     console.log('Dados do projeto:', project); // Verifique os dados do projeto
     setProjectData(project); // Define os dados do projeto selecionado
     navigate(`/kanban/${project.id}`, { state: { projectData: project } }); // Abra a tela Kanban para o projeto
   };
-
   return (
     <div className='table'>
       <div className='meus-projetos'>
